Improved performance of generated code with regard to dynamic binding. Currently we were using the   full dynamic type, but since tables are indexed with the dynamic type instead, it requires an   indirection to get the dtype from the dftype. This indirection costs about 4% on the compiler   speed and 2% in executable size. The solution was to store both dftype and dtype in the object   header. On 64-bit, the header is keeping its size, on 32-bit it goes from 8 bytes to 12 bytes.   That is to say on 32-bit, when you had 8MB of memory used for headers, it is now using 12MB,   and usually this corresponds to 70MB of data (for the compiler at least).
Consequence:
- dynamic types have a type definition EIF_TYPE_INDEX
- dynamic types are limited to 65000+ (before it was 32000+)
- updated runtime to use EIF_TYPE_INDEX wherever a dtype/dftype is used. Some places were using
  a int or long and this was abused (e.g. debugger encoding 2 dtypes on one int).

Other work:
- fixed most C warnings generated by the VS 2005 C++ comiler
- fixed some issues with equal/copy (use of dtype instead of dftype and comparison of expanded objects
  which should query the size in EIF_Size and not in the object header, since expanded on the stack
  have no size).
- fixed eweasel test#melt064.
- fixed eweasel test#exec255 which eventhough was passing, was not when enabling runtime assertions.
- fixed some issues with storables with respect to storable of special of expanded
- fixed issue with retrieval of special of expanded (see eweasel test#store013).
- simplified `st_write_func' to not take the flags since it is not needed for storing.

What's not perfect:
- There is a mix of usage of int, uint32 and int16 in the interpreter and in store/retrieve for representing
  types. It would be nice to make this more consistent and to use EIF_TYPE_INDEX instead.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@71783 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/C/run-time/include/rt_search.h
@@ -49,7 +49,7 @@ extern "C" {
 
 #define EIF_SEARCH_OK		0		/* No conflict result value of `s_put' */
 #define EIF_SEARCH_CONFLICT	1		/* Conflict result value of `s_put' */
-#define EIF_SEARCH_FOUND	-1		/* Object found by `s_search' */
+#define EIF_SEARCH_FOUND	(size_t) -1		/* Object found by `s_search' */
 
 /*
  * Search table interface.
